05-11

+++ TP OOP/tableaux/cc-services.php
@@ -1,5 +1,6 @@
 <?php
 session_start();
+include_once('../class/Service.php');
 
 if (!$_SESSION){
     header('Location: ../connexion.php');
@@ -16,7 +17,23 @@ if (!$_SESSION){
 </head>
 <body>
     
-    <a href="c-forms.php"><button type="button" class="btn btn-primary btn-lg" >Remplir le questionnaire</button></a>
+    
+    
+    <div class="boutons col-10 offset-1 mt-3 mb-3 row">
+        
+        <?php if ($_SESSION['profil']=="administrateur"){ ?>
+            <div class="  col-3 mb-1">
+                <a href="c-forms.php"><button type="button" class="btn btn-outline-primary " >Ajouter un service  +</button></a>
+            </div>
+        <?php } ?>
+        
+        <div class=" <?php if(isset($_SESSION) && $_SESSION['profil']=="administrateur"){ echo"col-6 mb-1";} else {echo "col-9 mb-1";} ?>">
+            <a href="employes.php" ><button name="employes" type="submit"  class="btn btn-primary "> EMPLOYES </button></a>
+        </div>
+        <div class=" col-3 mb-1">
+            <a href="../traitement.php?p=deco" ><button name="deconnexion" type="submit"  class="btn btn-outline-secondary "> Se déconnecter  X </button></a>
+        </div>
+    </div>
     <?php 
 
         include 'crud.php';    
@@ -26,7 +43,9 @@ if (!$_SESSION){
             if (isset($_GET['action']) && $_GET["action"]=="ajout" && !empty($_POST) &&
                 isset($_POST['noserv']) && !empty($_POST['noserv'])){
 
-                        add($_POST['noserv'], $_POST['serv'], $_POST['ville']);    
+                        $serv = new Service();
+                        $serv->setNoserv($_POST['noserv'])->setServ($_POST['serv'])->setVille($_POST['ville']);
+                        add($serv);    
                     }
             
             
@@ -41,8 +60,10 @@ if (!$_SESSION){
             //modifier
             elseif (isset($_GET["action"]) && $_GET["action"]=="modifier" && 
                     isset($_POST['noserv']) && !empty($_POST['noserv'])){
-
-                       edit($_GET['noserv'],$_POST['serv'], $_POST['ville'] );
+                        
+                        $serv = new Service();
+                        $serv->setNoserv($_GET['noserv'])->setServ($_POST['serv'])->setVille($_POST['ville']);
+                        edit($serv);
                         
                     }
                             ?>
@@ -55,8 +76,10 @@ if (!$_SESSION){
                 <th> N° service </th>
                 <th> service</th>
                 <th> ville </th>
-                <th> Supprimer</th>
+                <?php if ($_SESSION['profil']=="administrateur"){ ?>
+                <th> Supprimer </th>
                 <th> Modifier </th>
+                <?php } ?>
                 <th> Consulter </th>
             </tr>
         </thead>
@@ -69,6 +92,7 @@ if (!$_SESSION){
            if (isset ($_GET["action"]) && $_GET["action"]=="consulter" && 
                 isset($_GET['noserv']) && !empty($_GET['noserv'])){
 
+                    //CONSULTATION
                     $data=consult($_GET['noserv']);
 
                     foreach ($data as $key => $n){
@@ -83,46 +107,49 @@ if (!$_SESSION){
                     echo"<tr>";
 
                 foreach ($value as $k => $v) {
+                    
                         echo "<td>$v</td>";
+                  
                     }
-                
-                echo "<td>";
-                // fonction qui recense les services ayant des salariés
-                $donnees=tridelete();
-                $taille=count($donnees);
-                
-
-                $flag=false;
+                if ($_SESSION['profil']=="administrateur"){
+                    echo "<td>";
+                    // fonction qui recense les services ayant des salariés
+                    $donnees=tridelete();
+                    $taille=count($donnees);
                     
-                for ($i=0; $i<$taille; $i++) {
+
+                    $flag=false;
                         
-                    if ($value['noserv'] == $donnees[$i]['noserv']){
-                        $flag=true;
-                        break;
-                    } 
-                }
-
-                    if(!$flag){  
-                ?>
-                        <a href='cc-services.php?action=delete&amp;noserv=<?php echo $value['noserv']?>'>
-                        <button class='btn btn-outline-danger' value='Remove'>Supprimer</button>
-                        </a>  
-                <?php
-                    } 
-                    else {
-                ?>      <p> Ne peut être supprimé </p> <?php
+                    for ($i=0; $i<$taille; $i++) {
+                            
+                        if ($value['noserv'] == $donnees[$i]['noserv']){
+                            $flag=true;
+                            break;
+                        } 
                     }
-                ?>
                     
-                    </td>
-                    <td>
-                        <a href='c-forms.php?action=modifier&amp;noserv=<?php echo $value['noserv']?>'> 
-                            <button class='btn btn-outline-success' value='Modify'>Modifier</button>
-                            </a> 
-                    </td>
+                        if(!$flag){  
+                    ?>
+                            <a href='cc-services.php?action=delete&amp;noserv=<?php echo $value['noserv']?>'>
+                            <button class='btn btn-outline-danger' value='Remove'>Supprimer</button>
+                            </a>  
+                    <?php
+                        } 
+                        else {
+                    ?>      <p> Ne peut être supprimé </p> <?php
+                        }
+                    ?>
+                        
+                        </td>
+                        <td>
+                            <a href='c-forms.php?action=modifier&amp;noserv=<?php echo $value['noserv']?>'> 
+                                <button class='btn btn-outline-success' value='Modify'>Modifier</button>
+                                </a> 
+                        </td>
+                    <?php } ?>
                     <td>
                             <a href='cc-services.php?action=consulter&amp;noserv=<?php echo $value['noserv'] ?>'>
-                            <button class='btn btn-outline-info' value='Modify'>Consulter</button>
+                            <button class='btn btn-outline-info' value='Consult'>Consulter</button>
                             </a> 
                         </td>
                     </tr>
